+++ src/test/java/nextstep/test/NSTest.java
@@ -20,7 +20,7 @@ import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.Mockito.mockStatic;
 
 public abstract class NSTest {
-    private static final Duration TEST_TIMEOUT = Duration.ofSeconds(1L);
+    private static final Duration TEST_TIMEOUT = Duration.ofSeconds(5L);
 
     private PrintStream standardOut;
     private OutputStream captor;
